+++ test/shadcn_ui/components/forms/text_controls_test.exs
@@ -1,5 +1,5 @@
 defmodule ShadcnUI.Components.Forms.TextControlsTest do
-  use ExUnit.Case, async: true
+  use ExUnit.Case, async: false
 
   alias Phoenix.HTML.FormField
   alias Phoenix.HTML.Safe